requirements.txt
asgiref==3.2.10
astroid==2.4.2
backcall==0.2.0
colorama==0.4.3
decorator==4.4.2
Django==3.0.8
django-ckeditor==6.0.0
django-cleanup==5.2.0
django-crispy-forms==1.9.2
django-js-asset==1.2.2
django-recaptcha==2.0.6
ipython==7.16.1
ipython-genutils==0.2.0
isort==4.3.21
jedi==0.17.2
lazy-object-proxy==1.4.3
mccabe==0.6.1
parso==0.7.0
pexpect==4.8.0
pickleshare==0.7.5
Pillow==7.2.0
prompt-toolkit==3.0.5
psycopg2-binary==2.8.6
ptyprocess==0.7.0
Pygments==2.6.1
pylint==2.5.3
pytz==2020.1
six==1.15.0
sqlparse==0.3.1
toml==0.10.1
traitlets==4.3.3
wcwidth==0.2.5
wrapt==1.12.1
